requirements.txt
boto3==1.35.88
elevenlabs==2.7.1
numpy==2.3.1
openai==0.28.0
openai_whisper==20240930
playsound==1.2.2
PyAudio==0.2.14
pydub==0.25.1
python-dotenv==1.1.1
rich==14.0.0
